requirements.txt
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.1.4
click==8.1.8
Deprecated==1.3.1
distro==1.9.0
exceptiongroup==1.3.1
Flask==3.1.2
flask-cors==6.0.2
Flask-Limiter==3.11.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
limits==4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
openai==2.15.0
ordered-set==4.1.0
packaging==24.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
rich==13.9.4
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.5
wrapt==2.0.1
zipp==3.23.0
